Add conversions between terms and multidimensional arrays
This adds support for passing up to 3 dimensional arrays of Int64 or
Float64 in to the jl_call/N predicates, converting a nested Prolog list
to a Julia array. Any dimensional array of Int64 or Float64 can be
returned from jl_eval using the same Prolog form. NB the Prolog array
term is NOT yet recognised by the DCG, but this should be easy enough to
add next.